Idempotent POST /callbacks (#18)
* Return idempotent responses when identical messages are sent to POST /callbacks

* Write test cases for new methods, return JSON reponse with correct status code when Response is successful

* Cover all s3_callback_repo

* Create boto_client_responses.py for boto_client_response data, implement tests for response classes, remove redundant imports-codes

* Don't call data provider class again and again, remove contents key from list_objects' empty return

* Use data provider

* Call build() each time ResponseSuccess called

* convert pydantic model(ResponseSuccess) to JSONResponse in the endpoints

* Remove tuple_response, rename method to callback_exists(), cached fetched objects, decrease duplicated code using a new method named _get_object_from_s3

* Correct tests

* Cover new methods

* Improve __eq__ method using pydantic's dict() method

+++ app/repositories/s3_callback_repo.py
@@ -19,16 +19,42 @@ class S3CallbackRepo(CallbackRepo):
         super().__init__(**kwargs)
         with handle_s3_errors():
             self.s3 = boto3.client("s3", **settings.s3_configuration)
+        # Cached objects
+        self.callbacks = list()
+
+    def get_callback_from_cache(self, callback_dict: dict) -> Callback:
+        callback_obj = Callback(**callback_dict)
+        for callback in self.callbacks:
+            if callback == callback_obj:
+                return callback
+        # This line should be never reached
+        return  # noqa
+
+    def callback_exists(self, callback_dict: dict) -> bool:
+        callback_obj = Callback(**callback_dict)
+        with handle_s3_errors():
+            callbacks = self.s3.list_objects(Bucket=settings.CALLBACK_BUCKET)
+        if "Contents" not in callbacks:
+            return False
+
+        for row in callbacks["Contents"]:
+            callback = self._get_object_from_s3(row)
+            if callback == callback_obj:
+                return True
+
+        return False
 
     def save_callback(self, callback: dict) -> Callback:
-        cb = Callback(**callback)
-        with handle_s3_errors():
-            self.s3.put_object(
-                Body=bytes(cb.json(), "utf-8"),
-                Key=f"callbacks/{cb.enrolment_id}/{cb.callback_id}.json",
-                Bucket=settings.CALLBACK_BUCKET,
-            )
-
+        if self.callback_exists(callback):
+            cb = self.get_callback_from_cache(callback)
+        else:
+            cb = Callback(**callback)
+            with handle_s3_errors():
+                self.s3.put_object(
+                    Body=bytes(cb.json(), "utf-8"),
+                    Key=f"callbacks/{cb.enrolment_id}/{cb.callback_id}.json",
+                    Bucket=settings.CALLBACK_BUCKET,
+                )
         return cb
 
     def get_callbacks_list(self, enrolment_id: str):
@@ -45,12 +71,17 @@ class S3CallbackRepo(CallbackRepo):
             return {"callbacks_list": callbacks_list}
         # add callback_id and datetime in list
         for row in callbacks_objects_list["Contents"]:
-            with handle_s3_errors():
-                obj = self.s3.get_object(
-                    Key=row["Key"], Bucket=settings.CALLBACK_BUCKET
-                )
-            callback = Callback(**json.loads(obj["Body"].read().decode()))
+            callback = self._get_object_from_s3(row)
             callbacks_list.append(
                 {"callback_id": callback.callback_id, "received": callback.received}
             )
         return {"callbacks_list": callbacks_list}
+
+    def _get_object_from_s3(self, data):
+        with handle_s3_errors():
+            obj = self.s3.get_object(Key=data["Key"], Bucket=settings.CALLBACK_BUCKET)
+        callback = Callback(**json.loads(obj["Body"].read().decode()))
+        # Add fetched object to cache
+        if callback not in self.callbacks:
+            self.callbacks.append(callback)
+        return callback
